requirements.txt
pillow>=10.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
